requirements.txt
asgiref==3.8.1
Django==5.2.11
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.1
whitenoise==6.7.0
gunicorn==22.0.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
requests==2.32.3
